+++ src/helperFunctions/updateCardContent.ts
@@ -1,0 +1,25 @@
+import task from '@/interfaces/task.type';
+import { ToastTypes, toast } from './toast';
+import fetch from './fetch';
+import { TASKS_URL } from '@/components/constants/url';
+
+const { SUCCESS, ERROR } = ToastTypes;
+
+async function updateCardContent(id: string, cardDetails: task) {
+    try {
+        const { requestPromise } = fetch({
+            url: `${TASKS_URL}/${id}`,
+            method: 'patch',
+            data: cardDetails,
+        });
+        await requestPromise;
+        toast(SUCCESS, 'Changes have been saved !');
+    } catch (err: any) {
+        if ('response' in err) {
+            toast(ERROR, err.response.data.message);
+            return;
+        }
+        toast(ERROR, err.message);
+    }
+}
+export default updateCardContent;
